+++ include/view/view.h
@@ -19,7 +19,7 @@ void view_fatal(
 extern PyObject* ip_address;
 extern PyObject* invalid_status_error;
 extern PyObject* ws_handshake_error;
-extern PyObject* ws_closed;
+extern PyObject* route_log;
 
 #if defined(__LINE__) && defined(__FILE__)
 #define VIEW_FATAL(msg) view_fatal(msg, __FILE__, __func__, __LINE__)