+++ src/view/__init__.py
@@ -8,13 +8,10 @@ GitHub: https://github.com/zerointensity/view.py
 try:
     import _view
 except ModuleNotFoundError as e:
-    raise ImportError(
-        "_view has not been built, did you forget to compile it?"
-    ) from e
+    raise ImportError("_view has not been built, did you forget to compile it?") from e
 
 # these are re-exports
-from _view import (Context, InvalidStatusError, WebSocket,
-                   WebSocketHandshakeError)
+from _view import Context, InvalidStatusError, WebSocketHandshakeError
 
 from . import _codec
 from .__about__ import *