+++ src/view/_loader.py
@@ -5,8 +5,15 @@ import sys
 import warnings
 from dataclasses import _MISSING_TYPE, Field, dataclass
 from pathlib import Path
-from typing import (TYPE_CHECKING, ForwardRef, Iterable, NamedTuple, TypedDict,
-                    get_args, get_type_hints)
+from typing import (
+    TYPE_CHECKING,
+    ForwardRef,
+    Iterable,
+    NamedTuple,
+    TypedDict,
+    get_args,
+    get_type_hints,
+)
 
 from _view import Context
 
@@ -24,10 +31,13 @@ import inspect
 
 from ._logging import Internal
 from ._util import docs_hint, is_annotated, is_union, set_load
-from .exceptions import (DuplicateRouteError, InvalidBodyError,
-                         InvalidRouteError, LoaderWarning)
-from .routing import (BodyParam, Method, Route, RouteData, RouteInput,
-                      _NoDefault)
+from .exceptions import (
+    DuplicateRouteError,
+    InvalidBodyError,
+    InvalidRouteError,
+    LoaderWarning,
+)
+from .routing import BodyParam, Method, Route, RouteData, RouteInput, _NoDefault
 from .typing import Any, RouteInputDict, TypeInfo, ValueType
 
 ExtNotRequired = None
@@ -153,8 +163,7 @@ def _format_body(
         vbody_defaults[k] = default
 
     return [
-        (TYPECODE_CLASSTYPES, k, v, vbody_defaults[k])
-        for k, v in vbody_final.items()
+        (TYPECODE_CLASSTYPES, k, v, vbody_defaults[k]) for k, v in vbody_final.items()
     ]
 
 
@@ -344,9 +353,7 @@ def _build_type_codes(
                 vbody_types = vbody
 
             doc = {}
-            codes.append(
-                (TYPECODE_CLASS, tp, _format_body(vbody_types, doc, tp))
-            )
+            codes.append((TYPECODE_CLASS, tp, _format_body(vbody_types, doc, tp)))
             setattr(tp, "_view_doc", doc)
             continue
 
@@ -360,9 +367,7 @@ def _build_type_codes(
             key, value = get_args(tp)
 
             if key is not str:
-                raise InvalidBodyError(
-                    f"dictionary keys must be strings, not {key}"
-                )
+                raise InvalidBodyError(f"dictionary keys must be strings, not {key}")
 
             tp_codes = _build_type_codes((value,))
             codes.append((TYPECODE_DICT, None, tp_codes))
@@ -461,9 +466,7 @@ def finalize(routes: list[Route], app: ViewApp):
                     route.inputs.insert(index, 1)
                     continue
 
-                default = (
-                    v.default if v.default is not inspect._empty else _NoDefault
-                )
+                default = v.default if v.default is not inspect._empty else _NoDefault
 
                 route.inputs.insert(
                     index,
@@ -559,9 +562,7 @@ def load_fs(app: ViewApp, target_dir: Path) -> None:
                     )
                 else:
                     path_obj = Path(path)
-                    stripped = list(
-                        path_obj.parts[len(target_dir.parts) :]
-                    )  # noqa
+                    stripped = list(path_obj.parts[len(target_dir.parts) :])  # noqa
                     if stripped[-1] == "index.py":
                         stripped.pop(len(stripped) - 1)
 
@@ -614,8 +615,7 @@ def load_simple(app: ViewApp, target_dir: Path) -> None:
             for route in mini_routes:
                 if not route.path:
                     raise InvalidRouteError(
-                        "omitting path is only supported"
-                        " on filesystem loading",
+                        "omitting path is only supported" " on filesystem loading",
                     )
 
                 routes.append(route)