+++ src/view/_logging.py
@@ -4,7 +4,6 @@ import logging
 import os
 import queue
 import random
-import re
 import sys
 import time
 import warnings
@@ -26,12 +25,11 @@ from rich.progress_bar import ProgressBar
 from rich.table import Table
 from rich.text import Text
 
+from _view import setup_route_log
+
 from ._util import shell_hint
 from .exceptions import ViewInternalError
 from .typing import LogLevel
-
-UVICORN_ROUTE_REGEX = re.compile(r'.*"(.+) (\/.*) .+" ([0-9]{1,3}).*')
-
 
 # see https://github.com/Textualize/rich/issues/433
 
@@ -101,24 +99,7 @@ class UvicornHijack(logging.Filter):
     def filter(self, record: logging.LogRecord):
         if record.exc_text:
             Service.error(record.exc_text)
-        STATUS_TO_SVC = {
-            logging.DEBUG: Service.debug,
-            logging.INFO: Service.info,
-            logging.WARNING: Service.warning,
-            logging.ERROR: Service.error,
-            logging.CRITICAL: Service.critical,
-        }
-        message = record.getMessage()
-        match = UVICORN_ROUTE_REGEX.match(message)
-
-        if match:
-            # its a route message
-            method = match.group(1)
-            path = match.group(2)
-            status = int(match.group(3))
-            route(path, status, method)
-        else:
-            STATUS_TO_SVC[record.levelno](message)
+        Service.log.handle(record)
         return False
 
 
@@ -148,6 +129,7 @@ for lg in (svc, internal):
         show_path=False,
         show_time=False,
         rich_tracebacks=True,
+        markup=True,
     )
     handler.setFormatter(ViewFormatter())
     lg.addHandler(handler)
@@ -320,15 +302,15 @@ def _status_color(status: int) -> str:
 
 
 _METHOD_COLORS: dict[str, str] = {
-    "HEAD": "dim green",
-    "GET": "green",
-    "POST": "blue",
-    "PUT": "dim blue",
-    "PATCH": "cyan",
-    "DELETE": "red",
-    "CONNECT": "magenta",
-    "OPTIONS": "yellow",
-    "TRACE": "dim yellow",
+    "HEAD": "bold dim green",
+    "GET": "bold green",
+    "POST": "bold blue",
+    "PUT": "bold dim blue",
+    "PATCH": "bold cyan",
+    "DELETE": "bold red",
+    "CONNECT": "bold magenta",
+    "OPTIONS": "bold yellow",
+    "TRACE": "bold dim yellow",
 }
 
 
@@ -829,7 +811,7 @@ def _server_logger():
 
     try:
         import plotext as plt
-    except ModuleNotFoundError as e:
+    except ModuleNotFoundError:
         plt = None
 
     class Plot:
@@ -1023,6 +1005,21 @@ def _server_logger():
                     info.route,
                     f"[bold {_status_color(info.status)}]{info.status}[/]",
                 )
+
+
+def _write_route(status: int, route: str, method: str) -> None:
+    info = RouteInfo(status, route, method)
+
+    if _LIVE:
+        _QUEUE.put_nowait(QueueItem(True, True, "info", "", info))
+    else:
+        Service.info(
+            f"[{_METHOD_COLORS[method]}]{method.lower()}[/] [white]{route}[/] [{_status_color(status)}]{status}[/]",
+            highlight=False,
+        )
+
+
+setup_route_log(_write_route)
 
 
 def enter_server():