+++ src/view/routing.py
@@ -6,13 +6,11 @@ import re
 from contextlib import suppress
 from dataclasses import dataclass, field
 from enum import Enum
-from typing import (Any, Callable, Generic, Iterable, Literal, Type, TypeVar,
-                    Union)
+from typing import Any, Callable, Generic, Iterable, Literal, Type, TypeVar, Union
 
 from ._util import LoadChecker, make_hint
 from .exceptions import InvalidRouteError, MistakeError
-from .typing import (Middleware, StrMethod, Validator, ValueType, ViewResponse,
-                     ViewRoute)
+from .typing import Middleware, StrMethod, Validator, ValueType, ViewResponse, ViewRoute
 
 __all__ = (
     "get",
@@ -141,8 +139,7 @@ def route_types(
         route.extra_types[data.__name__] = data
     else:
         raise InvalidRouteError(
-            "expected type, tuple of tuples,"
-            f" or a dict, got {type(data).__name__}"
+            "expected type, tuple of tuples," f" or a dict, got {type(data).__name__}"
         )
 
     return route
@@ -167,17 +164,13 @@ def _method(
     if not util_path.startswith("/"):
         raise MistakeError(
             "paths must started with a slash",
-            hint=make_hint(
-                f'This should be "/{util_path}" instead', back_lines=2
-            ),
+            hint=make_hint(f'This should be "/{util_path}" instead', back_lines=2),
         )
 
     if util_path.endswith("/") and (len(util_path) != 1):
         raise MistakeError(
             "paths must not end with a slash",
-            hint=make_hint(
-                f'This should be "{util_path[:-1]}" instead', back_lines=2
-            ),
+            hint=make_hint(f'This should be "{util_path[:-1]}" instead', back_lines=2),
         )
 
     if "{" in util_path:
@@ -460,9 +453,7 @@ def route(
         doc,
         None,
         cache_rate,
-        method_list=[_STR_METHOD_MAPPING[i] for i in methods]
-        if methods
-        else None,
+        method_list=[_STR_METHOD_MAPPING[i] for i in methods] if methods else None,
     )