+++ tests/test_status.py
@@ -48,11 +48,7 @@ async def _():
 
     async with app.test() as test:
         for status in [*STATUS_CODES, *ERROR_CODES]:
-            assert (
-                await test.get("/", query={"status": status})
-            ).status == status
+            assert (await test.get("/", query={"status": status})).status == status
 
         for status in ERROR_CODES:
-            assert (
-                await test.get("/error", query={"status": status})
-            ).status == status
+            assert (await test.get("/error", query={"status": status})).status == status