+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 -   Added the `view docs` command
+-   Reworked internal logging API
+-   Added the `websocket` router
+-   Made hijack optional in fancy mode
 
 ## [1.0.0-alpha9] - 2024-2-4